Factor out JSON serialization. Store loanDetails.payments to local storage.
